feat(insc): embed ERC-7583 into ERC-721 and add Merkle verification
The receiving addresses and TokenIDs for the new INSC will be recorded in the server's database, and
this data will be organized to generate a Merkle tree. Each leaf node of the Merkle tree will
contain data about the owner and TokenID, ensuring that the new INSC is inscribed to the correct
address.