feat(transfer ft): integrate FT's transfer into this
Use a slot as the carrier for FT. Each user's first received NFT will be their Slot. The slot can
only be transferred at the end, ensuring that after the transfer, the user's balance of both NFT and
FT is zero. Once the FT function is activated, more slots can be minted. When a user conducts an FT
transaction and the recipient does not have a slot, a new slot without any balance will be minted to
carry the FT to be received. If the minting of slots reaches the limit, users can only purchase
slots from the NFT market to accept FT. A user's FT balance will only be the balance of the Slot
they hold. To use FT from other inscriptions, the waterToWine function must be used for conversion.