feat(passover): add a contract related to compensation
We will use an ERC20 token with a precision of 18 as our debt token, which will have a 1:1 ratio
with ETH. All users will be divided into three different batches for compensation. The first batch
includes users related to TokenIDs that have never been sold (due to theft or continuous holding).
The second batch is for users who have already sold their INSCs and are at the stage of making a
choice, namely the refund stage. The third batch is after the conclusion of the refund, followed by
compensation for the relevant users.